Use cmake instead of VS Solution, add comments and improvements to step2
