+++ step2/.gitignore
@@ -1,0 +1,3 @@
+build/
+compile_commands.json
+.cache/
