+++ step2/CMakeLists.txt
@@ -1,0 +1,9 @@
+cmake_minimum_required(VERSION 3.16)
+
+project(NinjaOneCppScreeningStep2)
+
+set(WINN32 true)
+
+set(CMAKE_CXX_STANDARD 20)
+
+add_executable(step2 CodeReview.cpp)
