--- step2/src/NinjaRRM_Screening_Step2.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.33529.622
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NinjaRRM_Screening_Step2", "NinjaRRM_Screening_Step2.vcxproj", "{A1997776-172E-45CD-999A-102789240DEC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A1997776-172E-45CD-999A-102789240DEC}.Debug|x64.ActiveCfg = Debug|x64
-		{A1997776-172E-45CD-999A-102789240DEC}.Debug|x64.Build.0 = Debug|x64
-		{A1997776-172E-45CD-999A-102789240DEC}.Debug|x86.ActiveCfg = Debug|Win32
-		{A1997776-172E-45CD-999A-102789240DEC}.Debug|x86.Build.0 = Debug|Win32
-		{A1997776-172E-45CD-999A-102789240DEC}.Release|x64.ActiveCfg = Release|x64
-		{A1997776-172E-45CD-999A-102789240DEC}.Release|x64.Build.0 = Release|x64
-		{A1997776-172E-45CD-999A-102789240DEC}.Release|x86.ActiveCfg = Release|Win32
-		{A1997776-172E-45CD-999A-102789240DEC}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {69613365-821E-4303-8087-F53698BD7593}
-	EndGlobalSection
-EndGlobal
